フォルダ構造少し変えて(コードの分割)、動作することを確認した。
