+++ CLAUDE.md
@@ -1,0 +1,50 @@
+## **1. 基本機能**
+
+**地図・位置情報**
+
+- ユーザーの現在位置を中心とした地図表示
+- 地図上での音声ピン(投稿ピン)の視覚的表示
+- 距離フィルター（10m以内のピンのみタップ可能）
+- 現在地ボタンによる地図の再センタリング
+
+**音声機能**
+
+- 音声録音機能（録音ボタンから遷移）
+- 録音した音声の再生・再録音
+- 音声付きピンの再生
+
+**ユーザー管理**
+
+- ユーザー登録・ログイン機能
+- プロフィール編集機能
+- アカウント設定画面
+
+### **2. コンテンツ管理**
+
+**投稿機能**
+
+- 音声投稿（位置情報付き）
+- タイトル設定（投稿時）
+- テキスト編集（100文字以内）
+
+**インタラクション機能**
+
+- 投稿へのいいね機能
+- 投稿へのコメント機能
+- コメントへのいいね機能
+
+### **3. レイヤー機能**
+
+**レイヤー管理**
+
+- レイヤー名設定
+- レイヤー説明（最大200文字）
+- レイヤーのカテゴリー設定（複数選択可）
+- レイヤーのガイドライン設定（最大200文字）
+- レイヤーの公開/非公開設定
+- レイヤーへの画像設定（任意1枚）
+
+**レイヤー利用**
+
+- レイヤー選択による投稿フィルタリング
+- レイヤーのお気に入り登録
