+++ src/features/audioPin/domain/entities/AudioPin.ts
@@ -1,0 +1,16 @@
+export interface AudioPin {
+  id: string;
+  title: string;
+  userName: string;
+  userImage: string;
+  audioUrl: any; // require() の戻り値を許可
+  description: string;
+  latitude: number;
+  longitude: number;
+}
+
+export interface AudioPlaybackState {
+  isPlaying: boolean;
+  currentTime: number;
+  duration: number;
+}
