+++ src/features/map/domain/entities/MapRegion.ts
@@ -1,0 +1,6 @@
+export interface MapRegion {
+  latitude: number;
+  longitude: number;
+  latitudeDelta: number;
+  longitudeDelta: number;
+}
